C++ common AutoInitializer.
