+++ test/Compatibility/Test_CompatibilitySetter_CS/Program.cs
@@ -10,6 +10,9 @@ class Program
         // Note: This test app currently only tests that the compatibility properties set in
         // the project file generates the ConfigureCompatibility() function and compiles successfully.
         Microsoft.Windows.ApplicationModel.WindowsAppRuntime.Compatibility.AutoInitialize.ConfigureCompatibility();
+        // App using the <WindowsAppSDKRuntimePatchMode*> or <WindowsAppSDKDisabledChanges> project properties should
+        // NOT directly call the above function. This test calls it just to verify that the function is generated.
+
         Console.WriteLine("hello world");
         Environment.Exit(0);
     }
